perf: duplicate deletion of perf event
a malicious app can open a perf event with constraint_duplicate
bit set, disable the event, and close the fd.  On closing the fd,
the perf_release() modification causes the kernel to clean up
the event as if it still were enabled, leading to the event
being removed from a list twice.

Change-Id: If761305d6fbb64913a17b4d56ece821978804b39
CRs-Fixed: 977563
Signed-off-by: Srinivasarao P <spathi@codeaurora.org>